requirements.txt
flask-graphql==1.4.1
graphene==1.4.1
graphene-sqlalchemy>=1.0
flask-sqlalchemy==2.2
flask-migrate==2.0.4
flask-shell-ipython==0.3.0